Revamped camera system
Previous rendering system was designed around the idea that the rendering would be set up by bounding the min and max values on both the X and Y axis.  This led to difficulty when the graph had to be moved on a pixel by pixel basis (rather than moved by graph units).

So the new camera system is now origin and zoom factor based, which allows for much finer grain of movement operations and simpler calculations.

This implementation does not yet support drawing points and segments yet, and it does not support alternative aspect ratios yet.